code/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-crawler"
version = "0.1.0"
description = "Self-hosted news crawler + AI summarizer + GitHub Pages publisher"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests",
  "feedparser",
  "trafilatura",
  "sqlalchemy",
  "psycopg2-binary",
  "tenacity",
  "PyGithub",
  "lxml",
  "openai",
  "python-dotenv",
]

[project.optional-dependencies]
test = ["pytest", "pytest-mock"]

[project.scripts]
news-crawler = "news_crawler.__main__:main"
news-crawler-ingest = "news_crawler.ingest:main"
news-crawler-publish = "news_crawler.publish:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["news_crawler*"]
